fix dtypes issue with compositing speed up
Creating the composited image can be sped up by only performing
calculations where there is a label to overlay. ndimage.find_objects
must have an input array with an int dtype, which is not guaranteed
in npz files (or necessary for other functionality, like actions).
Cast input array to int for find_objects call to get this to work
for float dtype label files without changing the content in the npz

+++ desktop/caliban.py
@@ -1407,7 +1407,7 @@ class CalibanWindow:
 
         # make whole composite image: only need to generate where there are labels
         if None in (dy1, dy2, dx1, dx2):
-            obj_list = ndimage.find_objects(current_ann)
+            obj_list = ndimage.find_objects(current_ann.astype(np.uint16))
             self.composite_view = np.copy(raw_RGB)
             for slice_obj in obj_list:
                 if slice_obj is not None:
